+++ doc/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 ======
 
+v0.10.12 - 05.03.2023
+----
+* Small fixes regarding iterating dictionaries;
+* New storage dict_with_game_state to collect data about game progress;
+* Added progress bar to queue window;
+* Changes in adding units to building queue and removing units from queue methods;
+* Units production process completed.
+
 v0.10.11 - 03.03.2023
 ----
 * Fixes in class Shop_unit_label;